feat: useNoteMessage, useNotificationMessage 커스텀훅 추가
